refactor(admin): extract Local Auth, NTLM, and Debug sections from PlatformFormEditor (#1781)
Follow-up to #1942/#2044/#2046/#2048/#2049/#2050/#2051/#2054/#2060/#2061/#2063: extracts
the three remaining self-contained blocks from PlatformFormEditor.jsx (Local Authentication
Settings, NTLM Authentication Settings, and Debug Settings) into
client/src/features/admin/components/platform-form/{LocalAuthSection,NtlmAuthSection,AuthDebugSection}.jsx.
PlatformFormEditor.jsx shrinks from 1,840 to 1,386 lines. Pure extraction, no behavior change.
